Add documentation for PEP 370 features in distutils (#10745).
Apart from adding a section to describe the user scheme, this changeset
also does some much needed cleanup:

- fixed inverted reST targets
- fixed some paths
- avoided duplicating the same options listing five or six times
- added missing entries for C headers locations
- added documentation for --install-lib
- fixed a few misuses of the option role (see #9312), but not all (not
  worth the time, but will do it in packaging docs)
- fixed some markup

The paths fixes were done with an eye on the source code in the install
command, so they really describe what?s actually done.  The situation on
Mac OS X is rather messy: the fix for #8084 touched site and sysconfig,
but distutils does not use these files anymore since the Great Revert.
I suspect we have a mismatched stdlib at the moment, and the fix is not
even clear (see discussion on #8084).