Slight cleanup in distutils test_dist.
I have tests to add in this file and it?s always nice to start from a
clean base.  I?ve also changed a test that used to write an invalid
config file ('[global]command_packages = etc.' on one line), but the
test passes before and after this change, so either it magically works
or the test is poorly written.  Sigh.